% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launch_application}
\alias{launch_application}
\title{Launch CytoDRAV.}
\usage{
launch_application(launch.browser = T, ...)
}
\arguments{
\item{launch.browser}{TRUE or FALSE. Launch default browser.}
}
\description{
Function requires no parameters but accepts optional parameters
that can be passed in \code{\link[shiny]{runApp}}, such as \code{launch.browser}
}
\examples{
launch_application()
launch_application(launch.browser=F)
}
